jobs-completion-notifier/requirements.txt
azure-identity
azure-mgmt-media
azure-servicebus
azure-eventgrid
python-dotenv
aiohttp
